+++ mockall/tests/automock_foreign_c.rs
@@ -37,3 +37,13 @@ fn returning() {
     ctx.expect().returning(i64::from);
     assert_eq!(42, unsafe{mock_ffi::foo(42)});
 }
+
+/// Ensure that the mock function can be called from C by casting it to a C
+/// function pointer.
+#[test]
+fn c_abi() {
+    let ctx = mock_ffi::foo_context();
+    ctx.expect().returning(i64::from);
+    let p: unsafe extern "C" fn(u32) -> i64 = mock_ffi::foo;
+    assert_eq!(42, unsafe{p(42)});
+}